Conflicts:
	inkubator-hrm/src/main/java/com/inkubator/hrm/web/employee/PlacementOfEmployeeWorkScheduleFormController.java
	inkubator-hrm/src/main/webapp/protected/employee/placement_of_employee_work_schedule.xhtml

+++ inkubator-hrm/src/main/java/com/inkubator/hrm/web/model/DistributionLeaveSchemeModel.java
@@ -20,7 +20,7 @@ public class DistributionLeaveSchemeModel implements Serializable{
     private Long golonganJabatanId;
     private Integer sortBy;
     private Integer orderBy;
-    private Integer startBalance;
+    private Double startBalance;
 
     public Long getLeaveSchemeId() {
         return leaveSchemeId;
@@ -94,11 +94,11 @@ public class DistributionLeaveSchemeModel implements Serializable{
         this.orderBy = orderBy;
     }
 
-    public Integer getStartBalance() {
+    public Double getStartBalance() {
         return startBalance;
     }
 
-    public void setStartBalance(Integer startBalance) {
+    public void setStartBalance(Double startBalance) {
         this.startBalance = startBalance;
     }
